referrer-policy/css-integration/external-import-stylesheet.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS integration - image from imported stylesheet (external)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/referrer-policy/generic/common.js"></script>
    <meta name="referrer" content="never">
  </head>
  <body>
    <p>Check that resources from imported stylesheets (loaded from external
    stylesheets) are loaded with the referrer and referrer policy from the
    external stylesheet.</p>

    <div class="styled"></div>

    <script>
      var css_test = async_test("Image from imported stylesheet (external).");
      var id = token();
      var crossOriginUrlPrefix = location.protocol + "//www1." + location.hostname + ":" + location.port;
      var cssUrl = crossOriginUrlPrefix + "/referrer-policy/generic/subresource/stylesheet.py?id=" + id + "&import-rule";
      var urlPrefix = location.protocol + "//" + location.hostname + ":" + location.port;
      var cssReferrer = urlPrefix + "/referrer-policy/generic/subresource/stylesheet.py?id=" + id;
      var imgUrl = urlPrefix + "/referrer-policy/generic/subresource/image.py" +
                   "?id=" + id + "&report-headers";

      var link = document.createElement("link");
      link.href = cssUrl;
      link.rel = "stylesheet";
      link.onload = function() {
        css_test.step_timeout(
            queryXhr.bind(this, imgUrl,
                          function(message) {
                            css_test.step(function() {
                              assert_own_property(message, "headers");
                              assert_own_property(message, "referrer");
                              assert_equals(message.referrer, cssReferrer);
                            });
                            css_test.done();
                          }),
            1000);
      };
      document.head.appendChild(link);
    </script>

    <div id="log"></div>
  </body>
</html>
